+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -198,6 +198,8 @@ public final class HubAdapter implements IHub {
       boolean bindToScope,
       @Nullable Date startTimestamp,
       boolean waitForChildren,
+      @Nullable Long idleTimeout,
+      boolean trimEnd,
       @Nullable TransactionFinishedCallback transactionFinishedCallback) {
     return Sentry.startTransaction(
         transactionContexts,
@@ -205,6 +207,8 @@ public final class HubAdapter implements IHub {
         bindToScope,
         startTimestamp,
         waitForChildren,
+        idleTimeout,
+        trimEnd,
         transactionFinishedCallback);
   }