+++ sentry/src/main/java/io/sentry/NoOpHub.java
@@ -154,6 +154,8 @@ public final class NoOpHub implements IHub {
       boolean bindToScope,
       @Nullable Date startTimestamp,
       boolean waitForChildren,
+      @Nullable Long idleTimeout,
+      boolean trimEnd,
       @Nullable TransactionFinishedCallback transactionFinishedCallback) {
     return NoOpTransaction.getInstance();
   }