+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -11,6 +11,7 @@ import io.sentry.transport.ITransportGate;
 import io.sentry.transport.NoOpEnvelopeCache;
 import io.sentry.transport.NoOpTransportGate;
 import io.sentry.util.Platform;
+import io.sentry.util.StringUtils;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -55,6 +56,9 @@ public class SentryOptions {
    * just not send any events.
    */
   private @Nullable String dsn;
+
+  /** dsnHash is used as a subfolder of cacheDirPath to isolate events when rotating DSNs */
+  private @Nullable String dsnHash;
 
   /**
    * Controls how many seconds to wait before shutting down. Sentry SDKs send events from a
@@ -296,9 +300,6 @@ public class SentryOptions {
   /** Control if profiling is enabled or not for transactions */
   private boolean profilingEnabled = false;
 
-  /** The cache dir. path for caching profiling traces */
-  private @Nullable String profilingTracesDirPath;
-
   /** Max trace file size in bytes. */
   private long maxTraceFileSize = 5 * 1024 * 1024;
 
@@ -312,6 +313,16 @@ public class SentryOptions {
 
   /** Proguard UUID. */
   private @Nullable String proguardUuid;
+
+  /**
+   * The idle time, measured in ms, to wait until the transaction will be finished. The transaction
+   * will use the end timestamp of the last finished span as the endtime for the transaction.
+   *
+   * <p>When set to {@code null} the transaction must be finished manually.
+   *
+   * <p>The default is 3 seconds.
+   */
+  private @Nullable Long idleTimeout = 3000L;
 
   /**
    * Contains a list of context tags names (for example from MDC) that are meant to be applied as
@@ -375,8 +386,10 @@ public class SentryOptions {
    *
    * @param dsn the DSN
    */
-  public void setDsn(@Nullable String dsn) {
+  public void setDsn(final @Nullable String dsn) {
     this.dsn = dsn;
+
+    dsnHash = StringUtils.calculateStringHash(this.dsn, logger);
   }
 
   /**
@@ -598,7 +611,11 @@ public class SentryOptions {
    * @return the cache dir. path or null if not set
    */
   public @Nullable String getCacheDirPath() {
-    return cacheDirPath;
+    if (cacheDirPath == null || cacheDirPath.isEmpty()) {
+      return null;
+    }
+
+    return dsnHash != null ? new File(cacheDirPath, dsnHash).getAbsolutePath() : cacheDirPath;
   }
 
   /**
@@ -607,10 +624,11 @@ public class SentryOptions {
    * @return the outbox path or null if not set
    */
   public @Nullable String getOutboxPath() {
-    if (cacheDirPath == null || cacheDirPath.isEmpty()) {
+    final String cacheDirPath = getCacheDirPath();
+    if (cacheDirPath == null) {
       return null;
     }
-    return cacheDirPath + File.separator + "outbox";
+    return new File(cacheDirPath, "outbox").getAbsolutePath();
   }
 
   /**
@@ -618,7 +636,7 @@ public class SentryOptions {
    *
    * @param cacheDirPath the cache dir. path
    */
-  public void setCacheDirPath(@Nullable String cacheDirPath) {
+  public void setCacheDirPath(final @Nullable String cacheDirPath) {
     this.cacheDirPath = cacheDirPath;
   }
 
@@ -1487,16 +1505,11 @@ public class SentryOptions {
    * @return the profiling traces dir. path or null if not set
    */
   public @Nullable String getProfilingTracesDirPath() {
-    return profilingTracesDirPath;
-  }
-
-  /**
-   * Sets the profiling traces dir. path
-   *
-   * @param profilingTracesDirPath the profiling traces dir. path
-   */
-  public void setProfilingTracesDirPath(@Nullable String profilingTracesDirPath) {
-    this.profilingTracesDirPath = profilingTracesDirPath;
+    final String cacheDirPath = getCacheDirPath();
+    if (cacheDirPath == null) {
+      return null;
+    }
+    return new File(cacheDirPath, "profiling_traces").getAbsolutePath();
   }
 
   /**
@@ -1551,6 +1564,24 @@ public class SentryOptions {
    */
   public void addContextTag(final @NotNull String contextTag) {
     this.contextTags.add(contextTag);
+  }
+
+  /**
+   * Returns the idle timeout.
+   *
+   * @return the idle timeout in millis or null.
+   */
+  public @Nullable Long getIdleTimeout() {
+    return idleTimeout;
+  }
+
+  /**
+   * Sets the idle timeout.
+   *
+   * @param idleTimeout the idle timeout in millis or null.
+   */
+  public void setIdleTimeout(final @Nullable Long idleTimeout) {
+    this.idleTimeout = idleTimeout;
   }
 
   /**
@@ -1743,6 +1774,9 @@ public class SentryOptions {
     }
     if (options.getProguardUuid() != null) {
       setProguardUuid(options.getProguardUuid());
+    }
+    if (options.getIdleTimeout() != null) {
+      setIdleTimeout(options.getIdleTimeout());
     }
   }