+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -26,6 +26,7 @@ import io.sentry.protocol.SentryTransaction
 import io.sentry.protocol.User
 import io.sentry.test.callMethod
 import io.sentry.util.HintUtils
+import io.sentry.util.StringUtils
 import java.io.File
 import java.nio.file.Files
 import java.util.Queue
@@ -1358,7 +1359,7 @@ class HubTest {
 
     @Test
     fun `isCrashedLastRun does not delete native marker if auto session is enabled`() {
-        val nativeMarker = File(file.absolutePath, EnvelopeCache.NATIVE_CRASH_MARKER_FILE)
+        val nativeMarker = File(hashedFolder(), EnvelopeCache.NATIVE_CRASH_MARKER_FILE)
         nativeMarker.mkdirs()
         nativeMarker.createNewFile()
         val hub = generateHub() as Hub
@@ -1369,7 +1370,7 @@ class HubTest {
 
     @Test
     fun `isCrashedLastRun deletes the native marker if auto session is disabled`() {
-        val nativeMarker = File(file.absolutePath, EnvelopeCache.NATIVE_CRASH_MARKER_FILE)
+        val nativeMarker = File(hashedFolder(), EnvelopeCache.NATIVE_CRASH_MARKER_FILE)
         nativeMarker.mkdirs()
         nativeMarker.createNewFile()
         val hub = generateHub {
@@ -1380,9 +1381,11 @@ class HubTest {
         assertFalse(nativeMarker.exists())
     }
 
+    private val dsnTest = "https://key@sentry.io/proj"
+
     private fun generateHub(optionsConfiguration: Sentry.OptionsConfiguration<SentryOptions>? = null): IHub {
         val options = SentryOptions().apply {
-            dsn = "https://key@sentry.io/proj"
+            dsn = dsnTest
             cacheDirPath = file.absolutePath
             setSerializer(mock())
             tracesSampleRate = 1.0
@@ -1401,5 +1404,11 @@ class HubTest {
         val mockClient = mock<ISentryClient>()
         sut.bindClient(mockClient)
         return Pair(sut, mockClient)
+    }
+
+    private fun hashedFolder(): String {
+        val hash = StringUtils.calculateStringHash(dsnTest, mock())
+        val fileHashFolder = File(file.absolutePath, hash!!)
+        return fileHashFolder.absolutePath
     }
 }